Safe-Termux v1.0.8

+++ source/uninstall.py
@@ -3,8 +3,10 @@ import shutil
 from utils import parse
 
 def uninstall() -> None:
-    shutil.rmtree(os.path.expanduser('~/.bloqueio'))
-    os.remove(os.path.expanduser('~/.hushlogin'))
+    shutil.rmtree(os.path.expanduser('~/.bloqueio'), ignore_errors=True)
+
+    if os.path.isfile(os.path.expanduser('~/.hushlogin')):
+        os.remove(os.path.expanduser('~/.hushlogin'))
 
     file = open("/data/data/com.termux/files/usr/etc/termux-login.sh", "r")
     data = file.read().splitlines()
